Move BootstrapMethodConfiguration to hosted package
(cherry picked from commit aa567235d8c2aac37668515edfe5ae1914e6f5ba)
